name-search/nta_api_use_batch/requirements.txt
google-cloud-storage
Pandas
gcsfs
Mojimoji
